+++ src/StateSmithTest/CodeFileWriter/Ex1.h
@@ -31,9 +31,6 @@ enum
 // forward declaration
 typedef struct Ex1 Ex1;
 
-// event handler type
-typedef void (*Ex1_Func)(Ex1* sm);
-
 // State machine constructor. Must be called before start or dispatch event functions. Not thread safe.
 void Ex1_ctor(Ex1* sm);
 
@@ -41,6 +38,7 @@ void Ex1_ctor(Ex1* sm);
 void Ex1_start(Ex1* sm);
 
 // Dispatches an event to the state machine. Not thread safe.
+// Note! This function assumes that the `event_id` parameter is valid.
 void Ex1_dispatch_event(Ex1* sm, Ex1_EventId event_id);