+++ src/StateSmithTest/Output/Gil/GilToJavaScriptTests.cs
@@ -5,7 +5,6 @@ using StateSmith.Output.UserConfig;
 using StateSmith.Output;
 using StateSmith.Output.Gil.JavaScript;
 using FluentAssertions;
-using StateSmith.Output.Gil;
 
 namespace StateSmithTest.Output.Gil;
 
@@ -14,7 +13,7 @@ public class GilToJavaScriptTests
     [Fact]
     public void Test()
     {
-        string gilCode = AlgoTests.BuildExampleGilFile(skipIndentation: false, out var sm).ToString();
+        string gilCode = GilFileTestHelper.BuildExampleGilFile(skipIndentation: false, out var sm).ToString();
 
         RenderConfigJavaScriptVars renderConfig = new()
         {