+++ src/StateSmithTest/Output/Sim/diagrams/PlantEx2.c
@@ -102,6 +102,7 @@ void PlantEx2_start(PlantEx2* sm)
 }
 
 // Dispatches an event to the state machine. Not thread safe.
+// Note! This function assumes that the `event_id` parameter is valid.
 void PlantEx2_dispatch_event(PlantEx2* sm, PlantEx2_EventId event_id)
 {
     switch (sm->state_id)