+++ src/StateSmithTest/Output/Sim/diagrams/PlantEx2.h
@@ -40,9 +40,6 @@ enum
 // forward declaration
 typedef struct PlantEx2 PlantEx2;
 
-// event handler type
-typedef void (*PlantEx2_Func)(PlantEx2* sm);
-
 // State machine constructor. Must be called before start or dispatch event functions. Not thread safe.
 void PlantEx2_ctor(PlantEx2* sm);
 
@@ -50,6 +47,7 @@ void PlantEx2_ctor(PlantEx2* sm);
 void PlantEx2_start(PlantEx2* sm);
 
 // Dispatches an event to the state machine. Not thread safe.
+// Note! This function assumes that the `event_id` parameter is valid.
 void PlantEx2_dispatch_event(PlantEx2* sm, PlantEx2_EventId event_id);
 
 // Thread safe.