+++ src/StateSmithTest/spec/1/c/Spec1Sm.c
@@ -96,6 +96,7 @@ void Spec1Sm_start(Spec1Sm* sm)
 }
 
 // Dispatches an event to the state machine. Not thread safe.
+// Note! This function assumes that the `event_id` parameter is valid.
 void Spec1Sm_dispatch_event(Spec1Sm* sm, Spec1Sm_EventId event_id)
 {
     switch (sm->state_id)