+++ src/StateSmithTest/spec/1/c/Spec1Sm.h
@@ -44,9 +44,6 @@ typedef struct Spec1Sm_Vars
 } Spec1Sm_Vars;
 
 
-// event handler type
-typedef void (*Spec1Sm_Func)(Spec1Sm* sm);
-
 // State machine constructor. Must be called before start or dispatch event functions. Not thread safe.
 void Spec1Sm_ctor(Spec1Sm* sm);
 
@@ -54,6 +51,7 @@ void Spec1Sm_ctor(Spec1Sm* sm);
 void Spec1Sm_start(Spec1Sm* sm);
 
 // Dispatches an event to the state machine. Not thread safe.
+// Note! This function assumes that the `event_id` parameter is valid.
 void Spec1Sm_dispatch_event(Spec1Sm* sm, Spec1Sm_EventId event_id);
 
 // Thread safe.